+++ src/functions.php
@@ -7,6 +7,7 @@
 require get_template_directory() . '/inc/customizer.php';
 require get_template_directory() . '/inc/template-tags.php';
 require get_template_directory() . '/inc/github-updater.php';
+require get_template_directory() . '/inc/mail-settings.php';
 require get_template_directory() . '/inc/qrcode-generator.php';
 require get_template_directory() . '/inc/shortcodes.php';